pyproject.toml
[project]
name = "fastapi-projects"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "black>=25.1.0",
    "fastapi[standard]>=0.115.12",
    "flake8>=7.2.0",
    "isort>=6.0.1",
    "npm>=0.1.1",
    "passlib>=1.7.4",
    "pre-commit>=4.2.0",
    "pyjwt>=2.10.1",
    "pytest>=8.4.1",
    "python-jose>=3.5.0",
    "requests>=2.32.4",
    "ruff>=0.11.13",
    "slowapi>=0.1.9",
    "sqlalchemy>=2.0.41",
]
